:art: FORMAT (reducers): move project row reducer to root
