+++ src/pages/Project/Row/index.ts
@@ -6,18 +6,6 @@ import ProjectRowVideoModel from "./models/ProjectRowVideoModel";
 import ProjectRowFileModel from "./models/ProjectRowFileModel";
 import ProjectRowTextModel, { TextStyle } from "./models/ProjectRowTextModel";
 
-import projectRowsReducer, {
-  projectRowsAdapter,
-} from "./services/projectRowsReducer";
-
-import {
-  selectProjectRowById,
-  selectProjectRowIds,
-  selectProjectRowEntities,
-  selectAllProjectRows,
-  selectTotalProjectRows,
-} from "./services/projectRowSelectors";
-
 export default ProjectRow;
 export {
   ProjectRowType,
@@ -27,11 +15,4 @@ export {
   ProjectRowImageModel,
   ProjectRowVideoModel,
   ProjectRowFileModel,
-  projectRowsReducer,
-  projectRowsAdapter,
-  selectProjectRowById,
-  selectProjectRowIds,
-  selectProjectRowEntities,
-  selectAllProjectRows,
-  selectTotalProjectRows,
 };
